Chatboo IA Whassap/api whast/requirements.txt
flask==2.3.3
openai==1.3.0
requests==2.31.0
mysql-connector-python==8.0.33
flask-cors==4.0.0